cp-431: * move logic of selectedJournal to sidebar
